refactor: refactor AuthorizationWithCodeScreen
