+++ src/view/pages/AuthorizeWithCodeScreen/style.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  view: {
+    flex: 1,
+    paddingHorizontal: 10,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  divider: {
+    marginVertical: 5,
+  },
+  textInput: {
+    width: '100%',
+    marginBottom: 5,
+  },
+})
